requirements.txt
Django>=3.0,<4.0
django-environ==0.8.1
gunicorn==20.1.0
authlib==1.0.0b2
bleach==4.1.0
boto3==1.17.105
whitenoise==5.3.0
drf-spectacular==0.21.0
celery==5.2.1
django-redis==5.1.0
psycopg2-binary
django-cors-headers>=3.2.1
hiredis>=2.0.0
six>=1.16.0
factory_boy>=3.2.1
firebase-admin
django-ninja
django-taggit
google-cloud-storage
google-cloud-secret-manager==2.12.0
stripe
django-storages[google]

pytest==6.2.5
pytest-asyncio>=0.14.0
pytest-django>=3.9.0

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-django
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-psycopg2